+++ R/r-to-pmml-define-function.R
@@ -46,10 +46,13 @@ define_function_get_pmml_string <- function(tokens, function_name) {
 
   # names of variables that are in the local scope of this function
   function_scope_variables <- c(function_param_name_tokens$text)
+  function_local_variables <- c()
   for(i in 1:nrow(top_level_function_body_expr_tokens)) {
     if(i != nrow(top_level_function_body_expr_tokens)) {
       inner_func_expr <- top_level_function_body_expr_tokens[i, ]
       inner_func_name <- define_function_get_inner_func_name(inner_func_expr, tokens, function_name)
+      var_name <- util_get_var_and_func_names(get_descendants_of_token(inner_func_expr, tokens))[1]
+      function_local_variables <- c(function_local_variables, var_name)
       pmml_function_string <- paste(
         pmml_function_string,
         get_pmml_string_for_expr_token_within_function(
@@ -125,6 +128,13 @@ define_function_get_pmml_string <- function(tokens, function_name) {
             pmml_string_for_r_code,
             pmml_string_for_return_arg_expr_token
           )
+          if(symbols_within_return_arg_expr_which_are_not_function_arguments[j, ]$text %in% function_local_variables) {
+            pmml_string_for_return_arg_expr_token <- gsub(
+              glue::glue('<TableLocator location="taxonomy" name="{symbols_within_return_arg_expr_which_are_not_function_arguments[j, ]$text}" />'),
+              glue::glue('<TableLocator>{pmml_string_for_r_code}</TableLocator>'),
+              pmml_string_for_return_arg_expr_token
+            )
+          }
         }
       }
 
@@ -341,10 +351,20 @@ define_function_get_pmml_str_for_expr <- function(
   )
 }
 
-define_function_get_pmml_str_for_token <- function(token, tokens, orig_func_name, orig_func_param_tokens, is_last_expr) {
+define_function_get_pmml_str_for_token <- function(
+  token, tokens, orig_func_name, orig_func_param_tokens, is_last_expr
+) {
   get_pmml_str_for_token <- pmml_generic_get_pmml_str_for_token(define_function_get_pmml_str_for_expr)
 
-  return(get_pmml_str_for_token(token, tokens, tokens_create_empty_tokens_df(), list(),  orig_func_name, orig_func_param_tokens, is_last_expr))
+  return(get_pmml_str_for_token(
+    token,
+    tokens,
+    tokens_create_empty_tokens_df(),
+    list(),
+    orig_func_name,
+    orig_func_param_tokens,
+    is_last_expr
+  ))
 }
 
 get_pmml_string_for_expr_token_within_function <- 
@@ -358,7 +378,15 @@ get_pmml_string_for_expr_token_within_function <-
              ) {
   pmml_string_for_initialization_expr_token <- ''
   if(if_expr_is(inner_function_expr_token, tokens)) {
-    pmml_string_for_initialization_expr_token <- define_function_get_pmml_str_for_expr(inner_function_expr_token, tokens, original_function_name, original_function_arg_tokens, FALSE, function_scope_variables)
+    pmml_string_for_initialization_expr_token <- 
+      define_function_get_pmml_str_for_expr(
+        inner_function_expr_token,
+        tokens,
+        original_function_name,
+        original_function_arg_tokens,
+        FALSE,
+        function_scope_variables
+      )
   } else {
     #Get the expression token which has the initialization code
     initialization_expr_token <- get_token_with_assignment_code(get_child_tokens_for_parent(inner_function_expr_token, tokens))
@@ -391,9 +419,9 @@ get_pmml_string_for_expr_token_within_function <-
       pmmlStringFor_function_call_for_current_symbol <- gsub(r_function_name, get_function_name_for_inner_function_expr_token(original_function_name, symbol_name), pmmlStringFor_function_call_for_current_symbol)
 
       pmml_string_for_initialization_expr_token <- gsub(glue::glue('<FieldRef field="{symbol_name}"/>'), pmmlStringFor_function_call_for_current_symbol, pmml_string_for_initialization_expr_token)
-        
+       
       pmml_string_for_initialization_expr_token <- gsub(
-        glue::glue('<TableLocator location="taxonomy" name="{symbol_name}"/>'),
+        glue::glue('<TableLocator location="taxonomy" name="{symbol_name}" />'),
         glue::glue('<TableLocator>{pmmlStringFor_function_call_for_current_symbol}</TableLocator>'),
         pmml_string_for_initialization_expr_token
       )